Refactored frontend architecture
/app/frontend/src
  /components
    /chat/
    ...

  /styles
    /chat
      ...
  /contexts
    PageState.context.tsx
    WebSocket.context.tsx

  /Chat
    /components
      ...
    Chat.viewModel.tsx
    Chat.model.tsx
    Chat.view.tsx
    Chat.context.tsx
