+++ app/frontend/src/Chat/components/Userlist.tsx
@@ -7,9 +7,9 @@ import {
   ListItemButton
 } from "@mui/material";
 import GroupIcon from "@mui/icons-material/Group";
-import ContextMenuUsers from "../../../components/ContextMenuUsers";
+import ContextMenuUsers from "../../components/ContextMenuUsers";
 import { UserStatus } from "kingpong-lib";
-import { ChatMemberRank } from "../ChatViewModel";
+import { ChatMemberRank } from "../Chat.viewModel";
 
 export enum ChatMemberStatus {
   OK = "OK",
