+++ app/frontend/src/Profile/Profile.viewModel.tsx
@@ -1,11 +1,11 @@
 import React, { createContext, useContext, useEffect } from "react";
 
 /** Module Imports */
-import { socket } from "src/contexts/WebSocketContext";
+import { socket } from "src/contexts/WebSocket.context";
 
 /** Temporary data type import */
 import { MatchHistoryItem, ProfileEntity, ProfileEvents } from "kingpong-lib";
-import { ProfileModelType, useProfileModel } from "./profile.model";
+import { ProfileModelType, useProfileModel } from "./Profile.model";
 
 /**
  * ProfileViewModelType
