+++ app/frontend/src/game/components/GameActionBar.tsx
@@ -1,13 +1,9 @@
 //This will return a name and ready status for a single player
 import * as React from "react";
-import {
-  Avatar,
-  ToggleButton,
-  Typography,
-} from "@mui/material";
+import { Avatar, ToggleButton, Typography } from "@mui/material";
 import Box from "@mui/material/Box";
 import CheckCircleIcon from "@mui/icons-material/CheckCircle";
-import { socket } from "src/contexts/WebSocketContext";
+import { socket } from "src/contexts/WebSocket.context";
 //If the player is the current user, it will have a toggleable button for ready status
 //If the player is the opposing user, it will have a badge that shows opponents ready status
