+++ app/frontend/src/index.tsx
@@ -3,14 +3,14 @@ import ReactDOM from "react-dom/client";
 import "./index.css";
 
 /** Providers */
-import { socket, WebSocketProvider } from "src/contexts/WebSocketContext";
+import { socket, WebSocketProvider } from "src/contexts/WebSocket.context";
 import { CssBaseline, ThemeProvider } from "@mui/material";
-import { PageStateProvider } from "./contexts/PageStateContext";
-import { ChatViewModelProvider } from "src/pages/chat/ChatViewModel";
-import { ProfileViewModelProvider } from "./views/profile/profile.viewModel";
+import { PageStateProvider } from "./contexts/PageState.context";
+import { ChatViewModelProvider } from "src/Chat/Chat.viewModel";
+import { ProfileViewModelProvider } from "./Profile/Profile.viewModel";
 
 import customTheme from "src/theme";
-import { RootView } from "./views/root.view";
+import { RootView } from "./root.view";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 const theme = customTheme();
