+++ app/frontend/src/styles/chat/ChatArea.css
